store_backend/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
Django==4.1.4
django-filter==22.1
django-phonenumber-field==7.0.1
django-phonenumbers==1.0.1
djangorestframework==3.14.0
h11==0.14.0
idna==3.4
Markdown==3.4.1
phonenumbers==8.13.2
Pillow==9.3.0
psycopg2-binary==2.9.5
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.13
uvicorn==0.20.0
